separated signup page from app.js
