+++ client/src/pages/Home.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+
+
+function Home(props) {
+
+    return(
+        <div>
+          <p1>Hello!</p1>
+        </div>
+
+    
+    );
+}
+
+
+export default Home;
