+++ package.json
@@ -21,8 +21,11 @@
   },
   "dependencies": {
     "axios": "^0.18.0",
+    "bootstrap": "^4.5.0",
     "express": "^4.16.3",
+    "if-env": "^1.0.4",
     "mongoose": "^5.3.16",
-    "if-env": "^1.0.4"
+    "react-bootstrap": "^1.1.1",
+    "react-router": "^5.2.0"
   }
 }
